chore(policy): move logic into policies_compute.go
This new file contains the logic for policies computation.
